feat/issue-crud (#29)
* Get Issue

* Init Create Issue

* Create Issue 2

* Create Issue 3

* Create Issue 4

* Update Issue

* Soft Delete Issue

* Update Issue 2

+++ src/Features/ProjectStatus/Entities/ProjectStatusEntity.cs
@@ -13,7 +13,7 @@ public class ProjectStatusEntity : BaseSoftDeletableEntity
     public string Name { get; set; }
 
     [Column("index")]
-    public int Index { get; set; }
+    public double Index { get; set; }
 
     [ForeignKey(nameof(Project))]
     [Column("project_id")]
